solver/requirements.txt
APScheduler==3.6.1
asn1crypto==0.24.0
bcrypt==3.1.7
capstone==4.0.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
filelock==3.0.12
Flask==1.1.1
idna==2.8
importlib-metadata==0.23
intervaltree==2.1.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==7.2.0
packaging==19.2
paramiko==2.6.0
pluggy==0.13.0
psutil==5.6.3
pwntools==2.2.1
py==1.8.0
pycparser==2.19
pyelftools==0.25
Pygments==2.4.2
PyNaCl==1.3.0
pypandoc==1.4
pyparsing==2.4.2
pyserial==3.4
PySocks==1.7.1
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
ROPGadget==5.8
six==1.12.0
sortedcontainers==1.5.10
toml==0.10.0
tox==3.14.0
tzlocal==2.0.0
unicorn==1.0.1
urllib3==1.25.4
virtualenv==16.7.5
Werkzeug==0.16.0
zipp==0.6.0